Update bech32 dependency to 0.8
Record concerns about bech32m as comments.
The Bech32 weakness is unlikely to affect our use of it for keys and
addresses, all of which expect fixed data length.
The changes for the general-purpose decoding procedure make it
accept both variants, but produce the original Bech32 from BIP-0173.
Address encoding and decoding is fixed to the original Bech32 to avoid
any ambiguity.